admin_services/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
Werkzeug
